Implement User Story 1
As a big fan of linguistics
So that I can look at different parts of sentences every day
I would like to be able to generate random nouns

(This required to get express and nightwatch to work before doing the feature tests. At the moment, it solves the user story in a very simple way and needs refactoring).